feat: successful sync room state across players
